Create seperate file to house all mockdata.
This should make it easier for future tests to share the same mock data.
